requirements.txt
requests
tqdm
loguru
python-dotenv